+++ app/javascript/application.js
@@ -1,1 +1,21 @@
-import './src/components/index'
+import "@hotwired/turbo-rails"	
+import "./controllers"
+
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import store from './src/redux/store';
+import App from './src/components/App';
+
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
+root.render(
+  <BrowserRouter>
+    <Provider store={store}>
+      <App />
+    </Provider>
+  </BrowserRouter>,
+);
+
