+++ webpack.config.js
@@ -26,13 +26,4 @@ module.exports = {
       maxChunks: 1,
     }),
   ],
-  resolve: {
-    fallback: {
-      dgram: false,
-      fs: false,
-      net: false,
-      tls: false,
-      child_process: false,
-    },
-  },
 };
